RNG-75: Moved NativeSeedType to a new class.
